+++ CHANGES.rst
@@ -29,6 +29,10 @@ Changes:
   with various files and listing of directory contents.
 - Allow ``colander.SchemaNode`` (with extensions for `OpenAPI` schema converters) to provide validation ``pattern``
   field directly with a compiled ``re.Pattern`` object.
+- Support `CWL` definition for ``cwltool:CUDARequirement`` to request the use of a GPU, including support for using
+  Docker with a GPU (resolves `#104 <https://github.com/crim-ca/weaver/issues/104>`_).
+- Support `CWL` definition for ``NetworkAccess`` to indicate whether a process requires outgoing IPv4/IPv6 network
+  access.
 
 Fixes:
 ------